requirements.txt
openai
anthropic
python-dotenv
requests
uuid